Solution Test 5:
To calculate the value of the missing angle, the program first must ensure that 2 of the 3 angles are provided. This is determined
by checking if only one of the triangle's angles is 0 (meaning it has not been input). Next, the program subtracts the sum of two angles
(with assigned) values from 180. The remainder is the assigned value emtpy angle.