Solution Test 6
First, a new static data member called "determineIfRigtAngle" was created.  Next,
all previous logic that was stored in the submitButton_click event was moved into other functions for more clarity. Next,
the label that calculates perimeter was changed to area. This was initially set to ensure that two sides were provided
which was determined to be a good halfway step.