Solution Test 7:
Wrote a method called calculateSides() that takes two arguments. The first is a string called "hypotenuse" which tells us which side of the triangle is the hypotenuse. The second parameter is a string called "emtpySide" which tells us which side (if any) of the triangle has not been provided. Using this, I'm able to isolate for the missing side (if there is any)